+++ src/sender/EventTriangleAPI.Sender.UnitTests/Helpers/SupportTicketOpenedEventHelper.cs
@@ -1,4 +1,4 @@
-using EventTriangleAPI.Shared.Domain.Entities;
+using EventTriangleAPI.Shared.Domain.Events;
 
 namespace EventTriangleAPI.Sender.UnitTests.Helpers;
