+++ src/shared/EventTriangleAPI.Shared.Domain/Events/TransactionRollBackedEvent.cs
@@ -1,7 +1,7 @@
-using EventTriangleAPI.Shared.Domain.Entities.Validation;
+using EventTriangleAPI.Shared.Domain.Events.Validation;
 using FluentValidation;
 
-namespace EventTriangleAPI.Shared.Domain.Entities;
+namespace EventTriangleAPI.Shared.Domain.Events;
 
 public class TransactionRollBackedEvent
 {
