+++ src/shared/EventTriangleAPI.Shared.Domain/Events/Validation/CreditCardDeletedEventValidator.cs
@@ -1,6 +1,6 @@
 using FluentValidation;
 
-namespace EventTriangleAPI.Shared.Domain.Entities.Validation;
+namespace EventTriangleAPI.Shared.Domain.Events.Validation;
 
 public class CreditCardDeletedEventValidator : AbstractValidator<CreditCardDeletedEvent>
 {
